last_update/requirements.txt
requests
python-telegram-bot
langdetect
textblob